+++ lib/twilio-ruby/rest/api/v2010/account/sip/domain.rb
@@ -130,8 +130,14 @@ module Twilio
               # @param [Boolean] sip_registration Whether to allow SIP Endpoints to register
               #   with the domain to receive calls. Can be `true` or `false`. `true` allows SIP
               #   Endpoints to register with the domain to receive calls, `false` does not.
+              # @param [Boolean] emergency_calling_enabled Whether emergency calling is enabled
+              #   for the domain. If enabled, allows emergency calls on the domain from phone
+              #   numbers with validated addresses.
+              # @param [Boolean] secure Whether secure SIP is enabled for the domain. If
+              #   enabled, TLS will be enforced and SRTP will be negotiated on all incoming calls
+              #   to this sip domain.
               # @return [DomainInstance] Newly created DomainInstance
-              def create(domain_name: nil, friendly_name: :unset, voice_url: :unset, voice_method: :unset, voice_fallback_url: :unset, voice_fallback_method: :unset, voice_status_callback_url: :unset, voice_status_callback_method: :unset, sip_registration: :unset)
+              def create(domain_name: nil, friendly_name: :unset, voice_url: :unset, voice_method: :unset, voice_fallback_url: :unset, voice_fallback_method: :unset, voice_status_callback_url: :unset, voice_status_callback_method: :unset, sip_registration: :unset, emergency_calling_enabled: :unset, secure: :unset)
                 data = Twilio::Values.of({
                     'DomainName' => domain_name,
                     'FriendlyName' => friendly_name,
@@ -142,6 +148,8 @@ module Twilio
                     'VoiceStatusCallbackUrl' => voice_status_callback_url,
                     'VoiceStatusCallbackMethod' => voice_status_callback_method,
                     'SipRegistration' => sip_registration,
+                    'EmergencyCallingEnabled' => emergency_calling_enabled,
+                    'Secure' => secure,
                 })
 
                 payload = @version.create(
@@ -247,8 +255,14 @@ module Twilio
               #   Endpoints to register with the domain to receive calls, `false` does not.
               # @param [String] domain_name The unique address you reserve on Twilio to which
               #   you route your SIP traffic. Domain names can contain letters, digits, and "-".
+              # @param [Boolean] emergency_calling_enabled Whether emergency calling is enabled
+              #   for the domain. If enabled, allows emergency calls on the domain from phone
+              #   numbers with validated addresses.
+              # @param [Boolean] secure Whether secure SIP is enabled for the domain. If
+              #   enabled, TLS will be enforced and SRTP will be negotiated on all incoming calls
+              #   to this sip domain.
               # @return [DomainInstance] Updated DomainInstance
-              def update(friendly_name: :unset, voice_fallback_method: :unset, voice_fallback_url: :unset, voice_method: :unset, voice_status_callback_method: :unset, voice_status_callback_url: :unset, voice_url: :unset, sip_registration: :unset, domain_name: :unset)
+              def update(friendly_name: :unset, voice_fallback_method: :unset, voice_fallback_url: :unset, voice_method: :unset, voice_status_callback_method: :unset, voice_status_callback_url: :unset, voice_url: :unset, sip_registration: :unset, domain_name: :unset, emergency_calling_enabled: :unset, secure: :unset)
                 data = Twilio::Values.of({
                     'FriendlyName' => friendly_name,
                     'VoiceFallbackMethod' => voice_fallback_method,
@@ -259,6 +273,8 @@ module Twilio
                     'VoiceUrl' => voice_url,
                     'SipRegistration' => sip_registration,
                     'DomainName' => domain_name,
+                    'EmergencyCallingEnabled' => emergency_calling_enabled,
+                    'Secure' => secure,
                 })
 
                 payload = @version.update(
@@ -389,6 +405,8 @@ module Twilio
                     'voice_url' => payload['voice_url'],
                     'subresource_uris' => payload['subresource_uris'],
                     'sip_registration' => payload['sip_registration'],
+                    'emergency_calling_enabled' => payload['emergency_calling_enabled'],
+                    'secure' => payload['secure'],
                 }
 
                 # Context
@@ -510,6 +528,18 @@ module Twilio
               end
 
               ##
+              # @return [Boolean] Whether emergency calling is enabled for the domain.
+              def emergency_calling_enabled
+                @properties['emergency_calling_enabled']
+              end
+
+              ##
+              # @return [Boolean] Whether secure SIP is enabled for the domain
+              def secure
+                @properties['secure']
+              end
+
+              ##
               # Fetch a DomainInstance
               # @return [DomainInstance] Fetched DomainInstance
               def fetch
@@ -536,8 +566,14 @@ module Twilio
               #   Endpoints to register with the domain to receive calls, `false` does not.
               # @param [String] domain_name The unique address you reserve on Twilio to which
               #   you route your SIP traffic. Domain names can contain letters, digits, and "-".
+              # @param [Boolean] emergency_calling_enabled Whether emergency calling is enabled
+              #   for the domain. If enabled, allows emergency calls on the domain from phone
+              #   numbers with validated addresses.
+              # @param [Boolean] secure Whether secure SIP is enabled for the domain. If
+              #   enabled, TLS will be enforced and SRTP will be negotiated on all incoming calls
+              #   to this sip domain.
               # @return [DomainInstance] Updated DomainInstance
-              def update(friendly_name: :unset, voice_fallback_method: :unset, voice_fallback_url: :unset, voice_method: :unset, voice_status_callback_method: :unset, voice_status_callback_url: :unset, voice_url: :unset, sip_registration: :unset, domain_name: :unset)
+              def update(friendly_name: :unset, voice_fallback_method: :unset, voice_fallback_url: :unset, voice_method: :unset, voice_status_callback_method: :unset, voice_status_callback_url: :unset, voice_url: :unset, sip_registration: :unset, domain_name: :unset, emergency_calling_enabled: :unset, secure: :unset)
                 context.update(
                     friendly_name: friendly_name,
                     voice_fallback_method: voice_fallback_method,
@@ -548,6 +584,8 @@ module Twilio
                     voice_url: voice_url,
                     sip_registration: sip_registration,
                     domain_name: domain_name,
+                    emergency_calling_enabled: emergency_calling_enabled,
+                    secure: secure,
                 )
               end